requirements.txt
streamlit
youtube-transcript-api
langchain
langchain-community
langchain-groq
langchain-core
python-dotenv
sentence-transformers
faiss-cpu